+++ packages/core/postgrest-js/test/relationships.test.ts
@@ -4,7 +4,7 @@ import { expectType, TypeEqual } from './types'
 import { z } from 'zod'
 import { RequiredDeep } from 'type-fest'
 
-const REST_URL = 'http://localhost:3000'
+const REST_URL = 'http://localhost:54321/rest/v1'
 const postgrest = new PostgrestClient<Database>(REST_URL)
 
 const UsersRowSchema = z.object({
@@ -41,7 +41,7 @@ test('nested query with selective fields', async () => {
           },
           {
             "id": 4,
-            "message": "Some message on channel wihtout details",
+            "message": "Some message on channel without details",
           },
         ],
         "username": "supabot",
@@ -99,7 +99,7 @@ test('nested query with multiple levels and selective fields', async () => {
               "slug": "other",
             },
             "id": 4,
-            "message": "Some message on channel wihtout details",
+            "message": "Some message on channel without details",
           },
         ],
         "username": "supabot",
@@ -225,7 +225,7 @@ test('one-to-many relationship', async () => {
             "channel_id": 3,
             "data": null,
             "id": 4,
-            "message": "Some message on channel wihtout details",
+            "message": "Some message on channel without details",
             "username": "supabot",
           },
         ],